// generated/src/app/api-models/cap-memb-util-hist.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapMembUtilHist {
  seqCapMembUtilHist: number;
  capModelId: string;
  firstContractMonth: Date;
  capEntityCode: string;
  seqCapPoolId: number;
  seqProvId: number;
  seqVendId: number;
  seqMembId: number;
  applyTo: string;
  capStoplossId: string;
  mostRecentMonth: Date;
  utilHistAmt: number;
}